refactor into get-hosts command
