+++ cmd/udm-pro-api-client/internal/app/app.go
@@ -2,8 +2,10 @@ package app
 
 import (
 	"github.com/jsumners/udm-pro-api-client/cmd/udm-pro-api-client/internal/config"
+	"github.com/jsumners/udm-pro-api-client/pkg/udm"
 )
 
 type CliApp struct {
 	Config *config.Configuration
+	Client *udm.UdmClient
 }
